fixed bug

+++ train_SPST.py
@@ -225,9 +225,9 @@ class DataLoad(Dataset):
 
         # split train to train part and validation part
         if partition == "train":
-            self.train_ind = np.asarray([i for i in range(self.num_examples) if i % 10 < 8]).astype(np.int)
+            self.train_ind = np.asarray([i for i in range(self.num_examples) if i % 10 < 8]).astype(np.int32)
             np.random.shuffle(self.train_ind)
-            self.val_ind = np.asarray([i for i in range(self.num_examples) if i % 10 >= 8]).astype(np.int)
+            self.val_ind = np.asarray([i for i in range(self.num_examples) if i % 10 >= 8]).astype(np.int32)
             np.random.shuffle(self.val_ind)
 
         io.cprint("number of " + partition + " examples in trgt_dataset : " + str(len(self.pc)))
